third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/imagecapture-origin-trial-interfaces-script-added.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ImageCapture - origin trial is enabled by script-added meta tag</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/origin-trials-helper.js"></script>
<script>
// Generate token with the command:
// generate_token.py  http://127.0.0.1:8000 ImageCapture --expire-timestamp=2000000000
var token = "AgnhzY0dhvMKsSPYNWHltCxTsLpVFzM/H0rZdpY4DeVo7ry4xm8VLRJU5aVLy+wSgnUa34pTTR017JFu4qnUAQwAAABUeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiSW1hZ2VDYXB0dXJlIiwgImV4cGlyeSI6IDIwMDAwMDAwMDB9";

let interfaces_to_check = ['ImageCapture','MediaSettingsRange','PhotoCapabilities'];

test(t => {
  OriginTrialsHelper.check_interfaces_missing(this, interfaces_to_check);
}, "ImageCapture related interfaces before adding trial token via script.");

OriginTrialsHelper.add_token(token);

test(t => {
  OriginTrialsHelper.check_interfaces(this, interfaces_to_check);
}, "ImageCapture related interfaces after adding trial token via script.");
</script>
