third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/imagecapture-origin-trial-interfaces-script-added-expected.txt

This is a testharness.js-based test.
FAIL ImageCapture related interfaces before adding trial token via script. assert_equals: Interface ImageCapture exists on provided object expected false but got true
PASS ImageCapture related interfaces after adding trial token via script. 
Harness: the test ran to completion.
